requirements.txt
gevent==22.10.2
requests==2.28.1
beautifulsoup4==4.11.1
IPy==1.1
dnspython==2.2.1
urllib3==1.26.14
aiohttp==3.8.1
aiohttp==3.8.1
dnsresolver==1.0.0
lxml==5.3.0
